refractor: integrate NB planner class and NB link statistics class, remove deprecated classes
